requirements.txt
bottlenose==1.1.7
Django==1.10
lxml==4.1.1
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.1
six==1.11.0
